### 2024-01-20 #6251 @editorialbot
[PRE REVIEW]: Benchmarking Hierarchical Reasoning with HierarchyCraft
**Submitting author:** @MathisFederico (<a href="http://orcid.org/0009-0002-2624-3417">Mathis Federico</a>)
**Repository:** https://github.com/IRLL/HierarchyCraft
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.4
