### 2024-01-22 56e3f30
🎨 Add add_edge_numbers option for html requirements graph
