+++ src/hcraft/requirements.py
@@ -74,12 +74,15 @@ width="90%"/>
 
 from enum import Enum
 from pathlib import Path
+import random
+from PIL import Image, ImageDraw, ImageFont
+
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Set, Tuple, Union
 
 import networkx as nx
 import numpy as np
 
-from matplotlib import colormaps
+import seaborn as sns
 from matplotlib import pyplot as plt
 import matplotlib.patches as mpatches
 from matplotlib.axes import Axes
@@ -89,6 +92,7 @@ from hebg.graph import draw_networkx_nodes_images, get_nodes_by_level
 from hebg.layouts.metabased import leveled_layout_energy
 
 from hcraft.render.utils import load_or_create_image, obj_image_path
+from hcraft.transformation import InventoryOperation, InventoryOwner
 
 if TYPE_CHECKING:
     from hcraft.elements import Item, Stack, Zone
@@ -135,7 +139,7 @@ class RequirementTheme:
     def __init__(
         self,
         default_color: Any = "black",
-        edge_colors=colormaps.get_cmap("Set1").colors,
+        edge_colors=None,
         **kwargs,
     ) -> None:
         self.colors = self.DEFAULT_COLORS.copy()
@@ -151,6 +155,9 @@ class RequirementTheme:
             )
             return f"#{hexes.upper()}"
 
+        if edge_colors is None:
+            edge_colors = sns.color_palette("colorblind")
+            random.shuffle(edge_colors)
         self.edges_colors = [rgba_to_hex(*color) for color in edge_colors]
 
     def color_node(self, node_type: RequirementNode) -> Any:
@@ -182,7 +189,7 @@ class Requirements:
     def draw(
         self,
         ax: Optional[Axes] = None,
-        theme: RequirementTheme = RequirementTheme(),
+        theme: Optional[RequirementTheme] = None,
         layout: "RequirementsGraphLayout" = "level",
         engine: DrawEngine = DrawEngine.PLT,
         save_path: Optional[Path] = None,
@@ -194,6 +201,11 @@ class Requirements:
             ax: Matplotlib Axes to draw on.
             layout: Drawing layout. Defaults to "level".
         """
+        if theme is None:
+            theme = RequirementTheme()
+
+        apply_color_theme(self.graph, theme)
+
         pos = compute_layout(self.digraph, layout=layout)
 
         if save_path:
@@ -566,9 +578,7 @@ class RequirementsGraphLayout(Enum):
     """Classic spring layout."""
 
 
-def apply_color_theme(
-    graph: nx.MultiDiGraph, theme: RequirementTheme = RequirementTheme()
-):
+def apply_color_theme(graph: nx.MultiDiGraph, theme: RequirementTheme):
     for node, node_type in graph.nodes(data="type"):
         graph.nodes[node]["color"] = theme.color_node(node_type)
         for pred, _, key in graph.in_edges(node, keys=True):
@@ -704,49 +714,9 @@ def _draw_html(
 ):
     from pyvis.network import Network
 
-    for node, node_obj in graph.nodes(data="obj"):
-        if node_obj is None:
-            continue
-        image_path = obj_image_path(node_obj, resources_path)
-        if image_path.exists():
-            graph.nodes[node]["shape"] = "image"
-            graph.nodes[node]["image"] = image_path.as_uri()
-
-    done_edges = []
-    for start, end, key, color in graph.edges(data="color", keys=True):
-        graph.edges[start, end, key]["hoverWidth"] = 0.1
-        graph.edges[start, end, key]["hoverWidth"] = 0.1
-        graph.edges[start, end, key]["arrows"] = {
-            "to": {"enabled": False},
[…]
