### 2024-01-22 5b3b588
✅ Fix optional dependencies in tests
