### 2024-01-22 1b3e0a5
📝 Update requirements graphs html paths for docs
