+++ src/hcraft/examples/__init__.py
@@ -4,40 +4,44 @@ Here is the table of available HierarchyCraft environments examples.
 
 If you built one of your own, send us a pull request so we can add it to the list!
 
-| Gym name                           | CLI name          | Reference                       |
-|:-----------------------------------|:------------------|:--------------------------------|
-| "MineHcraft-v1"                    | `minecraft`       | `hcraft.examples.minecraft`     |
-| "MiniHCraftEmpty-v1"               | `minicraft`       | `hcraft.examples.minicraft`     |
-| "MiniHCraftFourRooms-v1"           | `minicraft`       | `hcraft.examples.minicraft`     |
-| "MiniHCraftMultiRoom-v1"           | `minicraft`       | `hcraft.examples.minicraft`     |
-| "MiniHCraftCrossing-v1"            | `minicraft`       | `hcraft.examples.minicraft`     |
-| "MiniHCraftKeyCorridor-v1"         | `minicraft`       | `hcraft.examples.minicraft`     |
-| "MiniHCraftDoorKey-v1"             | `minicraft`       | `hcraft.examples.minicraft`     |
-| "MiniHCraftUnlock-v1"              | `minicraft`       | `hcraft.examples.minicraft`     |
-| "MiniHCraftUnlockPickup-v1"        | `minicraft`       | `hcraft.examples.minicraft`     |
-| "MiniHCraftBlockedUnlockPickup-v1" | `minicraft`       | `hcraft.examples.minicraft`     |
-| "TowerHcraft-v1"                   | `tower`           | `hcraft.examples.tower`         |
-| "RecursiveHcraft-v1"               | `recursive`       | `hcraft.examples.recursive`     |
-| "LightRecursiveHcraft-v1"          | `light-recursive` | `hcraft.examples.recursive`     |
-| "Treasure-v1"                      | `treasure`        | `hcraft.examples.treasure`      |
-| "RandomHcraft-v1"                  | `random`          | `hcraft.examples.random_simple` |
+| Gym name                         | CLI name          | Reference                                       |
+|:---------------------------------|:------------------|:------------------------------------------------|
+| MineHcraft-v1                    | `minecraft`       | `hcraft.examples.minecraft`                     |
+| MiniHCraftEmpty-v1               | `minicraft`       | `hcraft.examples.minicraft.empty`               |
+| MiniHCraftFourRooms-v1           | `minicraft`       | `hcraft.examples.minicraft.fourrooms`           |
+| MiniHCraftMultiRoom-v1           | `minicraft`       | `hcraft.examples.minicraft.multiroom`           |
+| MiniHCraftCrossing-v1            | `minicraft`       | `hcraft.examples.minicraft.crossing`            |
+| MiniHCraftKeyCorridor-v1         | `minicraft`       | `hcraft.examples.minicraft.keycorridor`         |
+| MiniHCraftDoorKey-v1             | `minicraft`       | `hcraft.examples.minicraft.doorkey`             |
+| MiniHCraftUnlock-v1              | `minicraft`       | `hcraft.examples.minicraft.unlock`              |
+| MiniHCraftUnlockPickup-v1        | `minicraft`       | `hcraft.examples.minicraft.unlockpickup`        |
+| MiniHCraftBlockedUnlockPickup-v1 | `minicraft`       | `hcraft.examples.minicraft.unlockpickupblocked` |
+| TowerHcraft-v1                   | `tower`           | `hcraft.examples.tower`                         |
+| RecursiveHcraft-v1               | `recursive`       | `hcraft.examples.recursive`                     |
[…]
