+++ src/hcraft/examples/tower.py
@@ -26,7 +26,9 @@ Each of the items 3, 4 and 5 requires items {0, 1, 2}.
 Each of the items 0, 1 and 2 requires nothing and can be crafted from the start.
 
 Requirements graph for H2-W3:
-![TowerHcraft example requirements graph](../../../docs/images/requirements_graphs/TowerHcraft-H2-W3.png)
+<div class="graph">
+.. include:: ../../../docs/images/requirements_graphs/TowerHcraft-H2-W3.html
+</div>
 
 """
 
