### 2024-01-23 20945a9
📝 Add web uri for html graphs
