### 2024-01-23 fdea5b2
📝 Fix MineHcraft requirements graph path
