### 2024-01-24 14dcb35
✅ 🔨 Make Transformation valid on only one or all zones Modify all tests accordigly Modify behaviors accordigly
