+++ src/hcraft/examples/minecraft/transformations.py
@@ -4,7 +4,9 @@ All used Minecraft recipies.
 
 """
 
-from typing import List
+from tkinter import W
+from typing import List, Optional
+from hcraft.elements import Item, Zone
 
 import hcraft.examples.minecraft.items as items
 from hcraft.examples.minecraft.tools import (
@@ -34,28 +36,31 @@ def build_minehcraft_transformations() -> List[Transformation]:
 
 
 def _move_to_zones() -> List[Transformation]:
-    name_prefix = "move-to-"
-    walk = [
-        #: Move to FOREST
-        Transformation(
-            name_prefix + zones.FOREST.name,
-            destination=zones.FOREST,
-            zones=[zone for zone in zones.OVERWORLD if zone != zones.FOREST]
-            + [zones.END],
-        ),
-        #: Move to SWAMP
-        Transformation(
-            name_prefix + zones.SWAMP.name,
-            destination=zones.SWAMP,
-            zones=[zone for zone in zones.OVERWORLD if zone != zones.SWAMP],
-        ),
-        #: Move to MEADOW
-        Transformation(
-            name_prefix + zones.MEADOW.name,
-            destination=zones.MEADOW,
-            zones=[zone for zone in zones.OVERWORLD if zone != zones.MEADOW],
-        ),
-    ]
+    def move_name(
+        destination: Zone,
+        from_zone: Optional[Zone] = None,
+        with_item: Optional[Item] = None,
+    ):
+        base_name = f"move-to-{destination.name}"
+        if from_zone is not None:
+            base_name += f"-from-{from_zone.name}"
+        if with_item is not None:
+            base_name += f"-with-{with_item.name}"
+        return base_name
+
+    #: Walk to surface zones of zones.OVERWORLD
+    walk = []
+    for destination in [zones.FOREST, zones.SWAMP, zones.MEADOW]:
+        for other_zone in zones.OVERWORLD:
+            if destination == other_zone:
+                continue
+            walk.append(
+                Transformation(
+                    move_name(destination, from_zone=other_zone),
+                    destination=destination,
+                    zone=other_zone,
+                )
+            )
 
     required_pickaxe_materials = {
         zones.UNDERGROUND: [Material.STONE, Material.IRON, Material.DIAMOND],
@@ -64,31 +69,31 @@ def _move_to_zones() -> List[Transformation]:
 
     #: Dig to zones.UNDERGROUND or zones.BEDROCK
     dig = []
-    for zone, materials in required_pickaxe_materials.items():
-        for material in materials:
-            pickaxe = MC_TOOLS_BY_TYPE_AND_MATERIAL[ToolType.PICKAXE][material]
-            other_zones = [
-                other_zone for other_zone in zones.OVERWORLD if other_zone != zone
-            ]
-            dig.append(
-                Transformation(
-                    name_prefix + f"{zone.name}-with-{pickaxe.name}",
-                    destination=zone,
-                    zones=other_zones,
-                    inventory_changes=[Use(PLAYER, pickaxe, consume=1)],
+    for destination, materials in required_pickaxe_materials.items():
+        for other_zone in zones.OVERWORLD:
+            if other_zone == destination:
+                continue
+            for material in materials:
+                pickaxe = MC_TOOLS_BY_TYPE_AND_MATERIAL[ToolType.PICKAXE][material]
+                dig.append(
+                    Transformation(
+                        move_name(destination, from_zone=other_zone, with_item=pickaxe),
+                        destination=destination,
+                        zone=other_zone,
+                        inventory_changes=[Use(PLAYER, pickaxe, consume=1)],
+                    )
                 )
-            )
 
     nether = [
-        #: Move to NETHER
+        #: Move to NETHER through portal
         Transformation(
-            name_prefix + zones.NETHER.name,
[…]
