### 2024-01-24 71c3603
🐛 Fix Transformation names were omitting zone when searching item
