### 2024-01-24 a2c2210
🐛 Fix duplicated search_for_item transformation in BlockedUnlockPickup
