+++ src/hcraft/planning.py
@@ -77,14 +77,20 @@ If multiple terminal groups exists in the purpose, the PDDL goal will be the hig
 
 from typing import TYPE_CHECKING, Dict, Optional, Union, List
 from copy import deepcopy
-import collections
 
+
+from hcraft.transformation import Transformation, InventoryOwner
+from hcraft.task import Task, GetItemTask, PlaceItemTask, GoToZoneTask
+from hcraft.purpose import Purpose
+from hcraft.elements import Zone, Item
 
 # unified_planning is an optional dependency.
+UPF_AVAILABLE = True
 try:
     import unified_planning.shortcuts as ups
     from unified_planning.plans import SequentialPlan
     from unified_planning.engines.results import PlanGenerationResult
+    from unified_planning.model.problem import Problem
 
     UserType = ups.UserType
     Object = ups.Object
@@ -92,39 +98,16 @@ try:
     IntType = ups.IntType
     Fluent = ups.Fluent
     InstantaneousAction = ups.InstantaneousAction
-    Problem = ups.Problem
     OneshotPlanner = ups.OneshotPlanner
     OR = ups.Or
     AND = ups.And
     GE = ups.GE
     LE = ups.LE
 
+
 except ImportError:
-    PlanGenerationResult = collections.namedtuple("PlanGenerationResult", "plan")
+    UPF_AVAILABLE = False
 
-    class Problem:
-        def __init__(self, name: str) -> None:
-            raise ImportError(
-                "Missing planning dependencies. Install with:\n"
-                "pip install hcraft[planning]"
-            )
-
-    UserType = collections.namedtuple("UserType", "")
-    Object = collections.namedtuple("Object", "")
-    BoolType = collections.namedtuple("BoolType", "")
-    IntType = collections.namedtuple("IntType", "")
-    Fluent = collections.namedtuple("Fluent", "")
-    InstantaneousAction = collections.namedtuple("InstantaneousAction", "")
-    OneshotPlanner = collections.namedtuple("OneshotPlanner", "")
-    OR = collections.namedtuple("Or", "")
-    AND = collections.namedtuple("And", "")
-    GE = collections.namedtuple("GE", "")
-    LE = collections.namedtuple("LE", "")
-
-from hcraft.transformation import Transformation, InventoryOwner
-from hcraft.task import Task, GetItemTask, PlaceItemTask, GoToZoneTask
-from hcraft.purpose import Purpose
-from hcraft.elements import Zone, Item
 
 if TYPE_CHECKING:
     from hcraft.state import HcraftState
@@ -141,6 +124,7 @@ class HcraftPlanningProblem:
         name: str,
         purpose: Optional["Purpose"],
         timeout: float = 60,
+        planner_name: Optional[str] = None,
     ) -> None:
         """Initialize a HierarchyCraft planning problem on the given state and purpose.
 
@@ -151,11 +135,17 @@ class HcraftPlanningProblem:
             timeout: Time budget (s) for the plan to be found before giving up.
                 Set to -1 for no limit. Defaults to 60.
         """
-        self.upf_problem: Optional[Problem] = self._init_problem(state, name, purpose)
+        if not UPF_AVAILABLE:
+            raise ImportError(
+                "Missing planning dependencies. Install with:\n"
+                "pip install hcraft[planning]"
+            )
+        self.upf_problem: "Problem" = self._init_problem(state, name, purpose)
         self.plan: Optional[SequentialPlan] = None
         self.plans: List[SequentialPlan] = []
         self.stats: List[Statistics] = []
         self.timeout = timeout
+        self.planner_name = planner_name
 
     def action_from_plan(self, state: "HcraftState") -> Optional[int]:
         """Get the next gym action from a given state.
[…]
