+++ tests/requirements/test_can_draw.py
@@ -48,4 +48,4 @@ def test_can_draw(env_class: Type[HcraftEnv], mocker: MockerFixture):
         filepath = requirements_dir / f"{env.name}.html"
         if not save:
             mocker.patch("pyvis.network.Network.write_html")
-        requirements.draw(engine="pyvis", save_path=filepath, with_web_uri=False)
+        requirements.draw(engine="pyvis", save_path=filepath, with_web_uri=True)
