### 2024-01-25 667156d
✨ Apply ruff
