### 2024-01-25 c4a4232
✏️ Fix some variable names spelling
