### 2024-01-25 0d0c74c
⬆️ 🐛 Upgrade UPF dependency in "all" optional requirements
