+++ pyproject.toml
@@ -27,7 +27,7 @@ gym = ["gym >= 0.26"]
 gui = ["pygame >= 2.1.0", "pygame-menu >= 4.3.8"]
 planning = ["unified_planning[aries,enhsp,lpg] >= 1.0.0"]
 htmlvis = ["pyvis<=0.3.1"]
-all = ["gym >= 0.26", "pygame >= 2.1.0", "pygame-menu >= 4.3.8", "unified_planning[enhsp] >= 0.5.0", "pyvis<=0.3.1"]
+all = ["gym >= 0.26", "pygame >= 2.1.0", "pygame-menu >= 4.3.8", "unified_planning[aries,enhsp,lpg] >= 1.0.0", "pyvis<=0.3.1"]
 dev = ["pytest", "pytest-cov", "pytest-mock", "pytest-check", "pytest-xdist", "pre-commit", "ruff"]
 
 [project.urls]
