### 2024-01-25 409810d
⬇️ 👷 Downgrade UPF version for CI
