### 2024-01-25 fc81b74
💥 👷 Remove Python 3.7 support for all dependencies Due to unified-planning with aries and/or lpg
