+++ src/hcraft/env.py
@@ -502,7 +502,11 @@ class HcraftEnv(Env):
         """
         return self.all_behaviors[task_to_behavior_name(task)]
 
-    def planning_problem(self, **kwargs) -> HcraftPlanningProblem:
+    def planning_problem(
+        self,
+        hierarchical: bool = False,
+        **kwargs,
+    ) -> HcraftPlanningProblem:
         """Build this hcraft environment planning problem.
 
         Returns:
@@ -530,7 +534,13 @@ class HcraftEnv(Env):
             assert env.purpose.is_terminated # Purpose is achieved
             ```
         """
-        return HcraftPlanningProblem(self.state, self.name, self.purpose, **kwargs)
+        return HcraftPlanningProblem(
+            self.state,
+            self.name,
+            self.purpose,
+            hierarchical=hierarchical,
+            **kwargs,
+        )
 
     def _step_output(self, reward: float, terminated: bool, truncated: bool):
         infos = {
