+++ src/hcraft/planning.py
@@ -82,7 +82,7 @@ from copy import deepcopy
 from hcraft.transformation import Transformation, InventoryOwner
 from hcraft.task import Task, GetItemTask, PlaceItemTask, GoToZoneTask
 from hcraft.purpose import Purpose
-from hcraft.elements import Zone, Item
+from hcraft.elements import Stack, Zone, Item
 
 # unified_planning is an optional dependency.
 UPF_AVAILABLE = True
@@ -91,6 +91,9 @@ try:
     from unified_planning.plans import SequentialPlan
     from unified_planning.engines.results import PlanGenerationResult
     from unified_planning.model.problem import Problem
+    from unified_planning.model.htn import HierarchicalProblem
+    from unified_planning.model.htn.task import Task as UpfHTask
+    from unified_planning.model.htn.method import Method
 
     UserType = ups.UserType
     Object = ups.Object
@@ -111,6 +114,7 @@ except ImportError:
 
 if TYPE_CHECKING:
     from hcraft.state import HcraftState
+    from hcraft.world import World
 
 Statistics = Dict[str, Union[int, float]]
 
@@ -123,6 +127,7 @@ class HcraftPlanningProblem:
         state: "HcraftState",
         name: str,
         purpose: Optional["Purpose"],
+        hierarchical: bool,
         timeout: float = 60,
         planner_name: Optional[str] = None,
     ) -> None:
@@ -140,7 +145,14 @@ class HcraftPlanningProblem:
                 "Missing planning dependencies. Install with:\n"
                 "pip install hcraft[planning]"
             )
-        self.upf_problem: "Problem" = self._init_problem(state, name, purpose)
+        if hierarchical:
+            self.upf_problem = self._init_hierarchical_problem(
+                state=state, name=name, purpose=purpose
+            )
+        else:
+            self.upf_problem = self._init_flat_problem(
+                state=state, name=name, purpose=purpose
+            )
         self.plan: Optional["SequentialPlan"] = None
         self.plans: List["SequentialPlan"] = []
         self.stats: List["Statistics"] = []
@@ -212,14 +224,19 @@ class HcraftPlanningProblem:
                 self.upf_problem, timeout=self.timeout
             )
         if results.plan is None:
-            raise ValueError("Not plan could be found for this problem.")
+            raise ValueError(
+                "Not plan could be found for this problem. Status: %s", results.status
+            )
         self.plan = deepcopy(results.plan)
         self.plans.append(deepcopy(results.plan))
         self.stats.append(_read_statistics(results))
         return results
 
-    def _init_problem(
-        self, state: "HcraftState", name: str, purpose: Optional["Purpose"]
+    def _init_flat_problem(
+        self,
+        state: "HcraftState",
+        name: str,
+        purpose: Optional["Purpose"],
     ) -> "Problem":
         """Build a unified planning problem from the given world and purpose.
 
@@ -233,28 +250,125 @@ class HcraftPlanningProblem:
             Problem: Unified planning problem.
         """
         upf_problem = Problem(name)
-        self.zone_type = UserType("zone")
-        self.player_item_type = UserType("player_item")
-        self.zone_item_type = UserType("zone_item")
+        self._add_base_flat_problem(upf_problem, state.world)
+        if purpose is not None and purpose.terminal_groups:
+            upf_problem.add_goal(self._purpose_to_goal(purpose))
+        self.update_problem_to_state(upf_problem, state)
+        return upf_problem
 
-        self.zones_obj: Dict[Zone, "Object"] = {}
-        for zone in state.world.zones:
-            self.zones_obj[zone] = Object(zone.name, self.zone_type)
+    def _init_hierarchical_problem(
+        self,
+        state: "HcraftState",
+        name: str,
+        purpose: Optional["Purpose"],
+    ):
[…]
