+++ tests/planning/test_mineHcraft.py
@@ -105,14 +105,6 @@ KNOWN_TO_FAIL_ITEM_FOR_PLANNER = {
     + GOLD_TIER_ITEMS
     + NETHER_LEVEL_ITEMS
     + END_LEVEL_ITEMS,
-    "lpg": WOOD_LEVEL_ITEMS
-    + STONE_LEVEL_ITEMS
-    + LEATHER_TIER_ITEMS
-    + IRON_TIER_ITEMS
-    + DIAMOND_TIER_ITEMS
-    + GOLD_TIER_ITEMS
-    + NETHER_LEVEL_ITEMS
-    + END_LEVEL_ITEMS,
 }
 
 
