### 2024-01-31 e3419bb
🎉 HEBG now solved all non-mineHcraft envs ! ! Requirements graph of BlockedUnlockPickup might be broken because of the two alternative transformations to specific zones
