### 2024-02-02 5b05e85
🎉 Solve all Hcraft envs with HEBG
