### 2024-02-08 b5e2408
🐛 Add graph precomputation to ensure call graph is well behaved
