+++ tests/solving_behaviors/test_doc_example.py
@@ -1,3 +1,4 @@
+from matplotlib import pyplot as plt
 import pytest
 import pytest_check as check
 
@@ -8,14 +9,27 @@ def test_doc_example():
     from hcraft.examples.minecraft.items import DIAMOND
     from hcraft.task import GetItemTask
 
+    draw_call_graph = True
+
+    if draw_call_graph:
+        _fig, ax = plt.subplots()
+
     get_diamond = GetItemTask(DIAMOND)
     env = MineHcraftEnv(purpose=get_diamond, max_step=50)
     solving_behavior = env.solving_behavior(get_diamond)
 
     done = False
     observation = env.reset()
+    env.render()
     while not done:
         action = solving_behavior(observation)
+
+        if draw_call_graph:
+            plt.cla()
+            solving_behavior.graph.call_graph.draw(ax)
+            plt.show(block=False)
+
         observation, _reward, done, _info = env.step(action)
+        env.render()
 
     check.is_true(get_diamond.terminated)  # DIAMOND has been obtained !
