### 2024-02-15 fda4805
✅ Solve first HDDL GetItem on dummy env with aries
