### 2024-02-19 8b0b826
🐛 🏷️ Fix typing undefined error
