### 2024-02-21 #6251 @logological
Thank you for your submission, @MathisFederico! I will be serving as its editor.  At this point I'd like to invite you to suggest to me any reviewers who may be qualified to review this submission.  You can [search our reviewer database](https://reviewers.joss.theoj.org/lookup) or nominate people you may know from your professional network or through familiarity with their work, but please do not nominate anyone who may have a conflict of interest.  I'll meanwhile compile my own list of candidate reviewers.
