### 2024-02-23 #6251 @logological
Well, in the interest of getting things started, I'll go ahead and approach some reviewers.  @lucaferranti, @lwu9, and @TimKam, would you be willing and able to review this JOSS submission?
