### 2024-02-23 #6251 @lwu9
Hi Tristan,

I'm interested in reviewing this work. Thanks

Best,
Lili

On Fri, 23 Feb 2024 at 16:15, Tristan Miller ***@***.***>
wrote:

> Well, in the interest of getting things started, I'll go ahead and
> approach some reviewers. @lucaferranti <https://github.com/lucaferranti>,
> @lwu9 <https://github.com/lwu9>, and @TimKam <https://github.com/TimKam>,
> would you be willing and able to review this JOSS submission?
>
