### 2024-02-27 #6251 @logological
@nathanael-fijalkow and @jorgedch, would you be willing and able to review this JOSS submission?
