### 2024-03-12 97bf01f
👷 Remove Java from no-deps CI
