### 2024-03-12 48e1662
⬆️ Bump aries to fix internal error
