### 2024-03-27 #6251 @cst0
> @seba-1511 and @cst0, would you be willing and able to review this JOSS submission?

Hi @logological, sorry for my delay (I've been away from internet for the last few weeks)-- is there further action for me to take here? Happy to help if reviewers are still needed.
