### 2024-04-15 #6468 @lwu9
1. The there are two authors in the paper but there are three contributors in GitHub.
2. I was able to install the package, but when I ran "action = get_human_action(env)", it got stuck.
3. I didn't find clear community guidelines.
