### 2024-04-28 #6468 @Christopher-Henry-UM
I have performed most of the review except I am not comfortable checking the _Functionality_ or the _Performance_ boxes. The software is used to create arbitrary hierarchical environments that are compatible with the OpenAI Gym Reinforcement Learning framework or the AIPlan4EU Unified Planning Framework. This library enables users to easily create complex hierarchical structures that can be used to test and develop various reinforcement learning or planning algorithms.

While I am familiar with the concept of reinforcement learning, I do not have the experience with the type of programming this library was created for. As a result, I have no way to test it other than to carry out the Quickstart steps listed in the documentation.
