### 2024-05-01 #6468 @MathisFederico
Thanks for the reviews so far !

Sorry @logological for being long to answer, it has been two very busy weeks...

## About @lwu9 issues:
> 1. There are two authors in the paper but there are three contributors in GitHub.

I was not sure if I could add @SWang848 and @liyuxuan-academic because most of their contributions where non-code ones (testing, reviewing documentation, discussing ideas, ...) but re-reading [JOSS authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) it seems that I could indeed ! I asked them if they want to be added as author (and not just being in acknowledgments as now) and will update according to their answers.

>     2. I was able to install the package, but when I ran "action = get_human_action(env)", it got stuck.

When using human actions, the program will pause and wait for the action of the human, meaning you !
A pygame window should pop (be sure to have gui optional requirements as indicated in the readme):

![Initial pygame window for the readme example)](https://github.com/openjournals/joss-reviews/assets/60117466/462fd80c-11f7-4652-9ec7-55f3ab79231d)

You can select action be either using arrows and pressing enter, or by just clicking on available actions.
Selecting wood block would result in this:

![image](https://github.com/openjournals/joss-reviews/assets/60117466/75d7886e-19c4-4c77-b536-fc43ac748f74)

>     3. I didn't find clear community guidelines.

There was a CONTRIBUTING.md file but the readme didn't mention it indeed ! It is now added !

## About @Christopher-Henry-UM issues:

There are tests that checks those facts automatically both for planning (see [planning tests](https://github.com/IRLL/HierarchyCraft/tree/master/tests/planning)) and for [gym](https://github.com/IRLL/HierarchyCraft/blob/master/tests/test_gym.py) also for the gym compatibility this library has been heavily used with [stable-baselines3](https://github.com/DLR-RM/stable-baselines3) that requires the gym API to be valid.

I still asked around in the planning community to know if someone would like to review, I will update here if I find someone !
