### 2024-05-09 d7b0f4b
📝 Fix paper formatting
