### 2024-05-11 #6468 @MathisFederico
I agree it would be better but its hard to convince people to spend a bit of time on this, I'm still on the lookout too
