### 2024-05-27 #6468 @logological
> I agree it would be better but its hard to convince people to spend a bit of time on this, I'm still on the lookout too

I've contacted a few more potential reviewers and will of course keep everyone apprised.
