### 2024-05-27 #6468 @lwu9
Sounds good, thanks for the heads up. I will finish the check list today.

On Sun, May 26, 2024 at 11:43 PM Tristan Miller ***@***.***>
wrote:

> I agree it would be better but its hard to convince people to spend a bit
> of time on this, I'm still on the lookout too
>
> I've contacted a few more potential reviewers and will of course keep
> everyone apprised.
>
