### 2024-05-28 #6468 @lwu9
My previous questions have been resolved, thanks for the clarification. When I ran the snippets in Readme, I ran into two issues.
1. 'from hcraft.examples import TowerHcraft, RecursiveHcraft, RandomHcraft'
  It gave me import errors, but if I changed it to 'from hcraft.examples import TowerHcraftEnv, RecursiveHcraftEnv, RandomHcraftEnv', it worked.

2. 'problem.solve()' in the last section
  It gave me 'FileNotFoundError' in [\lib\site-packages\unified_planning\engines\pddl_planner.py:194], in PDDLPlanner._solve(self, problem, heuristic, timeout, output_stream)
