### 2024-06-05 #6468 @lwu9
I posted some comments on the github last week, but they didn't show up in
this email thread, so I just copied them here just in case:

My previous questions have been resolved, thanks for the clarification,
Mathïs! When I ran the snippets in Readme, I ran into two more issues.

   1.

   'from hcraft.examples import TowerHcraft, RecursiveHcraft, RandomHcraft'
   It gave me import errors, but if I changed it to 'from hcraft.examples
   import TowerHcraftEnv, RecursiveHcraftEnv, RandomHcraftEnv', it worked.
   2.

   'problem.solve()' in the last section
   It gave me 'FileNotFoundError' in
   [\lib\site-packages\unified_planning\engines\pddl_planner.py:194], in
   PDDLPlanner._solve(self, problem, heuristic, timeout, output_stream)

On Tue, 4 Jun 2024 at 15:29, Tristan Miller ***@***.***>
wrote:

> @yg211 <https://github.com/yg211> Hope all is well with you! Would you be
> willing and able to review this submission for the Journal of Open Source
> Software? We're particularly looking for a reviewer who has familiarity
> with the OpenAI Gym or AIPlan4EU platforms.
>
