### 2024-06-22 #6468 @MathisFederico
>     1. 'from hcraft.examples import TowerHcraft, RecursiveHcraft, RandomHcraft'
>        It gave me import errors, but if I changed it to 'from hcraft.examples import TowerHcraftEnv, RecursiveHcraftEnv, RandomHcraftEnv', it worked.

Indeed thanks for spotting it, fixed now.

>     2. 'problem.solve()' in the last section
>        It gave me 'FileNotFoundError' in [\lib\site-packages\unified_planning\engines\pddl_planner.py:194], in PDDLPlanner._solve(self, problem, heuristic, timeout, output_stream)

I can't seem to reproduce that error, may have more informations about your system and the pip list result for package versions ?
