### 2024-06-22 e81a925
📝 Fix README example envnames
