### 2024-07-08 #6468 @maltehelmert
> @maltehelmert @roeger Would you be willing and able to review this submission for the Journal of Open Source Software? We're particularly looking for a reviewer who has familiarity with the OpenAI Gym or AIPlan4EU platforms.

I'm sorry I'm currently overcommitted. :-( I think it's the same for @roeger. But perhaps you can recommend someone from the AIPlan4EU contributors, Gabi?
