### 2024-07-15 #6468 @logological
@alvalentini @arbimo @mikand Would you be willing and able to review this submission for the Journal of Open Source Software? We're particularly looking for a reviewer who has familiarity with the OpenAI Gym or AIPlan4EU platforms.
