### 2024-08-03 #6468 @MathisFederico
> @MathisFederico Thanks for your continued patience.  I have been continuing to reach out to potential reviewers by e-mail, and will see if I can contact a few more today through GitHub.  Hopefully we will get someone to accept soon.

I'm not in a hurry no problem, Thanks for doing what you do.

I have to fix the AI4Plan compatibility with their last breaking changes anyway and adding the Hierarchical Task Network conversion is something that I have to finish up.

I guess software is always evolving anyway and I intend to maintain this repo as long as the question "How do we quantify the hierarchicality of a behavior, task and environment" remains open, and it is still wide open today.
