### 2024-08-05 #6468 @inpefess
> @inpefess @lutzhamel Would you be willing and able to review this submission for the Journal of Open Source Software? We're particularly looking for a reviewer who has familiarity with the OpenAI Gym or AIPlan4EU platforms.

Yes, it would be a pleasure for me to review this submission. I don't know anything about AIPlan4EU yet, but I have relveant experience in creating and using OpenAI Gym environments.
