### 2024-08-05 #6468 @madmage
Dear Tristan Miller, is there any deadline for this review? I would like to
participate, but I would not have time in the next weeks.

On Sat, 3 Aug 2024 at 23:59, Tristan Miller ***@***.***>
wrote:

> @madmage <https://github.com/madmage> Would you be willing and able to
> review this submission for the Journal of Open Source Software? We're
> particularly looking for a reviewer who has familiarity with the OpenAI Gym
> or AIPlan4EU platforms.
>
