### 2024-08-08 #6468 @logological
@madmage It seems that another potential reviewer gave a positive response to my call a couple hours before you did, and so we wouldn't need your review after all.  Nonetheless thanks for getting back to me, and I hope you might consider reviewing for JOSS in the future!
