### 2024-08-08 #6468 @logological
@inpefess I've added you as a reviewer, but it seems I may have been mistaken about being able to generate the review checklist for you.  I think that you will need to do this yourself: please post a comment in this issue with the text `@editorialbot generate my checklist`.
