### 2024-08-09 4564798
👷 Remove 'all' optional requirements keyword to avoid duplication
