+++ .github/workflows/python-tests-all-optdeps.yml
@@ -31,11 +31,6 @@ jobs:
       with:
         distribution: "microsoft"
         java-version: "17"
-    - name: Download and install ENHSP
-      run: |
-        git clone https://gitlab.com/enricos83/ENHSP-Public.git
-        cd ENHSP-Public; git checkout enhsp20-0.9.7; bash ./compile; cd ..
-        bash -c "mkdir .planners; mv ENHSP-Public/enhsp-dist .planners/enhsp-20; rm -rf ENHSP-Public"
     - name: Test with pytest
       run: |
         pytest tests
