### 2024-08-11 IRLL/HierarchyCraft#43 @inpefess
Complete planning solving example
Part of a JOSS review https://github.com/openjournals/joss-reviews/issues/6468 There is an example of solving a planning problem in ReadMe (https://github.com/IRLL/HierarchyCraft?tab=readme-ov-file#as-a-upf-problem-for-planning) but it's incomplete. For example, when I try adding proposed lines:

```Python
problem = env.planning_problem()
problem.solve()
```

to Minecraft example (https://github.com/IRLL/HierarchyCraft?tab=readme-ov-file#using-the-programmatic-interface) nothing really happens. Moreover, if I try using the plan as suggested in the documentation https://irll.github.io/HierarchyCraft/hcraft/planning.html, it fails with an error (probably because no plan was built):

```
SequentialPlan:
Traceback (most recent call last):
  File "hcraft-review/venv/lib/python3.12/site-packages/hcraft/env.py", line 402, in step
    action = int(action)
             ^^^^^^^^^^^
TypeError: int() argument must be a string, a bytes-like object or a real number, not 'NoneType'
```

Could you please complete the example to get started with solving the planning problem based on Minecraft environment?
