### 2024-08-11 IRLL/HierarchyCraft#44 @inpefess
Comunity guidelines
Part of a JOSS review https://github.com/openjournals/joss-reviews/issues/6468
According to JOSS requirements, there should be "clear guidelines" to "2) Report issues or problems with the software 3) Seek support". I think that for GitHub users it's obvious that one should post issues (like this one) to report problems or get support, but could you please mention this explicitly in ReadMe for potential users who are somewhat new to GitHub workflow?
