### 2024-08-12 IRLL/HierarchyCraft#46 @inpefess
Update environment creating tutorial
One of the claims of the JOSS paper https://github.com/openjournals/joss-reviews/issues/6468#issuecomment-1989414549 is that one can add new environments to HierarchyCraft. I tried to follow the tutorial https://irll.github.io/HierarchyCraft/hcraft/env.html and it's really impressive, but seems a bit outdated:

1.  `from hcraftnv import HcraftEnv` --- there is no `hcraftnv` module (I assumed `hcraft.env`)
2. `from hcraft.elements import world_from_transformations` --- there is no `world_from_transformations` in `elements` (but it exists in `hcraft.world`)
3. https://pixel-poem.itch.io/dungeon-assetpuck looks cool, but seems to be paid

So, unfortunately, I was unable to finish the tutorial. Could you please update it?
