### 2024-08-12 #6468 @inpefess
@MathisFederico Thank you for developing HierarchyCraft! It's really cool! I added several issues related to my check-list to your repository (linked to this one), please have a look at them when you have time.
