### 2024-08-15 #6468 @MathisFederico
@inpefess Thank you for your review ! Lots of very good issues, I'm currently in the process of making my research packages uploaded on nixpkgs for better reproducibility.

I'll take care of all the issues you mentioned right after that !
