### 2024-08-19 IRLL/HierarchyCraft#44 @MathisFederico
There is a [Contributing](https://github.com/IRLL/HierarchyCraft?tab=readme-ov-file#contributing) section in the readme that redirects to [CONTRIBUTION.md](https://github.com/IRLL/HierarchyCraft/blob/master/CONTRIBUTING.md).

@inpefess  Where would you want to move it to be more visible ? (as you couldn't find it I guess it is misplaced)
