### 2024-08-19 4a977ef
✅ Test warn planning problem without goal
