### 2024-08-19 9997583
🔧 Update pre-commit for pytest on nix
