### 2024-08-19 5a8934b
📝 Improve planning doc example
