### 2024-08-19 b161183
📝 Make examples doc more clear and exaustive
