### 2024-08-19 IRLL/HierarchyCraft#43 @MathisFederico
By default, the MineHcraft environment has no goal, as such any plan built will be trivially empty, this is why the action given was None.

I've added a UserWarning if building a planning problem from and environment without goal.

I've updated the documentation to make everything more clear both in the [README](https://github.com/IRLL/HierarchyCraft?tab=readme-ov-file#as-a-upf-problem-for-planning) and in the [planning documentation](https://irll.github.io/HierarchyCraft/hcraft/planning.html) itself.

If you really want to use the MineHCraft example, you can either [set a purpose manually](https://irll.github.io/HierarchyCraft/hcraft/purpose.html) or use gym to make an env with predetermined purpose, for example:
```python
import gym

from hcraft.env import HcraftEnv

env: HcraftEnv = gym.make("MineHcraft-Stone-v1")

# Make it into a unified planning problem
planning_problem = env.planning_problem()

# Solve the planning problem and show the plan
planning_problem.solve()
print(planning_problem.plan)

done = False
_observation = env.reset()
while not done:
    action = planning_problem.action_from_plan(env.state)
    if action is None:
        # Plan has ended, no action left
        done = True
        continue
    _observation, _reward, done, _ = env.step(action)

# Goal is achieved == purpose is terminated
if env.purpose.terminated:
    print("Success ! The plan worked in the actual environment !")
else:
    print("Failed ... Something when wrong with the plan.")
```

@inpefess tell me if you are satisfied and thus if the issue should be closed
