### 2024-08-19 fcbfb69
🐛 📝 Fix env.py example documentation
