### 2024-08-19 IRLL/HierarchyCraft#46 @MathisFederico
Thank you for trying it !

1. Indeed there was some minor name changes, I've updated the tutorial accordingly !
2. Same as 1
3. The price is free for assets, 0$ included ;) --- But that is not even required as it is now hopefully [better shown in the documentation](https://irll.github.io/HierarchyCraft/hcraft/env.html#add-graphics) !

Could you please tell me if you could successfully finish the tutorial with the updates ? Thanks @inpefess !
