### 2024-08-19 #6468 @MathisFederico
@inpefess I've (hopefully) fixed the issues you raised code-wise, I'll get to the paper issue in the following week probably.
Could you please tell me for each issue if you confirm the fix ? Thanks !
