### 2024-08-21 IRLL/HierarchyCraft#46 @inpefess
Amazing! I didn't notice this small "just take me to the downloads" link, sorry. Now it works for me, I can customize images and even solve a plan for this game by appying other parts of the documentation. Thank you!
