+++ paper.bib
@@ -232,3 +232,42 @@
   url = 	 {https://proceedings.mlr.press/v176/hambro22a.html},
   abstract = 	 {In this report, we summarize the takeaways from the first NeurIPS 2021 NetHack Challenge. Participants were tasked with developing a program or agent that can win (i.e., ’ascend’ in) the popular dungeon-crawler game of NetHack by interacting with the NetHack Learning Environment (NLE), a scalable, procedurally generated, and challenging Gym environment for reinforcement learning (RL). The challenge showcased community-driven progress in AI with many diverse approaches significantly beating the previously best results on NetHack. Furthermore, it served as a direct comparison between neural (e.g., deep RL) and symbolic AI, as well as hybrid systems, demonstrating that on NetHack symbolic bots currently outperform deep RL by a large margin. Lastly, no agent got close to winning the game, illustrating NetHack’s suitability as a long-term benchmark for AI research.}
 }
+
+@article{botvinick2014model,
+  title={Model-based hierarchical reinforcement learning and human action control},
+  author={Botvinick, Matthew and Weinstein, Aaron},
+  journal={Philosophical Transactions of the Royal Society B: Biological Sciences},
+  volume={369},
+  number={1655},
+  pages={20130480},
+  year={2014},
+  publisher={The Royal Society}
+}
+
+@inproceedings{bacon2017option,
+  title={The option-critic architecture},
+  author={Bacon, Pierre-Luc and Harb, Mohamad and Precup, Doina},
+  booktitle={Proceedings of the AAAI Conference on Artificial Intelligence},
+  volume={31},
+  number={1},
+  year={2017}
+}
+
+@inproceedings{heess2016learning,
+  title={Learning continuous control policies by stochastic value gradients},
+  author={Heess, Nicolas and Wayne, Greg and Silver, David and Lillicrap, Timothy and Erez, Tom and Tassa, Yuval},
+  booktitle={Advances in Neural Information Processing Systems (NeurIPS)},
+  volume={29},
+  pages={2944--2952},
+  year={2016}
+}
+@inproceedings{nachum2018data,
+  title={Data-efficient hierarchical reinforcement learning},
+  author={Nachum, Ofir and Gu, Shixiang and Lee, Honglak and Levine, Sergey},
+  booktitle={Advances in Neural Information Processing Systems (NeurIPS)},
+  volume={31},
+  pages={3303--3313},
+  year={2018}
+}
+
+
