### 2024-08-31 6cb738e
🐛 Fix use of list for typing
