### 2024-08-31 805c671
🐛 Fix test not skipped when upf missing
