### 2024-09-10 #6468 @logological
@MathisFederico I see that all the issues raised by @inpefess except for IRLL/HierarchyCraft#45 are now closed, and that a commit referencing the latter issue was applied a few days ago, though the commit message doesn't speak to every comment raised in that issue.  Will you be continuing to revise the paper to address the remaining comments in that issue?  (Not that I am trying to rush you—I'm just trying to keep track of the progress so that it doesn't stall.)
