### 2024-09-14 #6468 @MathisFederico
Yes I will, sorry this past month has been filled with a lot of new things (new place, new job, ...)
But I do intend to finish it (the paper part is not my strong point)
