### 2025-01-25 #6468 @MathisFederico
Hey @logological ! We finally finished all the modifications following up on @inpefess and are ready to resume the review process. Thanks to @inpefess, the paper is much clearer now, thanks a lot for the detailed feedback !

Also we made a new release (2.0.0) that move the compatibility from [OpenAI Gym](https://github.com/openai/gym) (the old interface used by the RL community but not maintained anymore) to [Gymnasium](https://github.com/Farama-Foundation/Gymnasium) that is now maintained by [Farama Foundation](https://github.com/Farama-Foundation) (along with [minigrid](https://github.com/Farama-Foundation/MiniGrid) that is cited in the paper).
