### 2025-03-10 #6468 @inpefess
@MathisFederico sorry for prolonged silence and thank you for the new release! The text became much more solid than before, it was a pleasure to read. The last thing - please fix OpenAI Gym citation on line 138.

@logological I've finished my review, and after the mentioned minor correction, I recommend publication.
