### 2025-03-11 #6468 @MathisFederico
Thank you for reviewing @inpefess !

I have fixed the OpenAI gym reference that was both broken and not up to date as we now are compatible with the more recent gymnasium framework instead since OpenAI gym is not maintained anymore. Thanks for catching it !
