### 2025-03-12 #6468 @Christopher-Henry-UM
Hi Tristen,

Sure, no problem!

Please give me a timeline. Things are pretty busy with the hiring process, but I should be able to fit it in if you give me a deadline.

Cheers,

Chris

Get Outlook for Android<https://aka.ms/AAb9ysg>
