### 2025-03-28 #6468 @logological
Unfortunately I have not heard back from @lwu9 but if Chris manages to complete the _Community guidelines_ item of the review, then among all three reviewers every item in the reviewer checklist will have been completed by at least two reviewers.
