### 2025-04-03 #6468 @Christopher-Henry-UM
Hi Tristan, Sorry for the delay in responding. I am confident in checking the Community guidelines box. The authors have provided clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
