### 2025-04-06 #6468 @logological
@MathisFederico We've now got all the review criteria confirmed by 2 out of 3 reviewers, so I'm ready to do the final checks necessary before recommending acceptance.  As it's the end of the semester here the upcoming week is very busy for me, so I might have to wait until the week after that to start.  But once I do start, I don't think the process will take long.
