### 2025-04-16 #6468 @logological
@MathisFederico Thanks for your patience.  I've read through the text and identified a number of small errors in the language, formatting, and references.  Could you please address these?

* Title: Change "builder For" to "Builder for".
* Page 1: Change "we excludes" to "we exclude".
* Figure 1: Many of the text labels are too small to read.  Could you increase the size of the figure to the full width of the text block?
* Figure 1 caption: Change "Reinforcement", "Planning", "Hierarchical", and "Program" to "reinforcement", "planning", "hierarchical", and "program".
* Figure 2 caption: Change "There is" to "There are".
* Page 3: Consider changing "An exemplary instance" to "an example" or "a good example".  (Though the word "exemplary" has a literal sense of "serving as an example", it is most commonly used in the sense of "deserving honour, admiration, or imitation", and so sounds awkward here.)
* Page 3: Change "leaderboard(Milani" to "leaderboard (Milani".
* Page 3: Consider adding a comma before "and poses similar challenges" so that the reader does not mistakenly attach this independent clause to the "with".
* Page 3: This same independent clause looks to be a run-on sentence, since it contains two main verbs ("poses" and "rewards").  That is, "rewards sparsity and long-term reasoning" is not a noun phrase like the other items in the list, so if this is meant to be another example of a "similar challenge", you'll have to reword this accordingly.  If it's not an example of a "similar challenge" but rather a consequence of those challenges, or separate feature of Crafter, then it would be best to split this part off into a new sentence.
* Figure 3 caption: Change "Inspired from Figure 4 of (Hafner, 2002)" to "Inspired by Figure 4 of Hafner (2002)".
* Page 4: Change "This omission" to "These omissions".
* Page 4: Change "Each Atari games" to "Each Atari game".
* Page 4: Change "as each task ehibiting" to "with each task exhibiting" or "as each task exhibits".
* Page 4: Change "Gridworld, this makes" to "Gridworld; this makes".
* Page 4: Change "more wide that deep" to "more wide than deep".
* Page 4: Change "independently from" to "independently of".
* Page 4: Change "HierarchyCraft motivation" to "HierarchyCraft's motivation".
* Pages 5 and 6, including Figure 4: Inconsistent capitalization of "Requirements Graph" vs. "Requirements graph".  Come to think of it, do you even need to capitalize and italicize "Requirements [G|g]raph" and "Transformations"?  Why not leave these terms in lowercase (except at the beginning of a sentence or figure label) and in an upright font?  (I see that Transformation is a class name in the Python source, but "Requirements Graph" is not a valid Python class name, and so I assume the paper text isn't actually referring to real class names.)
* Page 5: Change "But each of" to "But each of the".
* Page 5: Change "require the further features extraction" to "require the further extraction of features".
* Page 6: Change "allow the use" to "allows the use".
* Page 6: Change "list of Transformation" to "list of Transformations".
* Page 7: Change "there assistance" to "their assistance".
* Throughout: The paper mixes American and Commonwealth English.  If you want American English, then change "parametrised" to "parameterized".  If you prefer Commonwealth English, then change the "z" to "s" in "utilized", "emphasize", "generalization", and "customize", and change "focusing" to "focussing".
* Throughout: Inconsistent use of "sub-task" vs. "subtask".  Choose one spelling and apply it consistently.
* All figures: For accessibility purposes, please consider replacing the bitmap images with scalable vector images, such as PDFs, that encode text as text.  (This isn't absolutely necessary, but there's a good chance that you created these figures in a vector drawing program to begin with, and so substituting vector versions should be fairly low effort.)
* References: Wherever possible, please provide DOIs and page numbers for the bibliography items, and check for and correct/remove any incorrect data from them.  For example, Bacon et al. (2017) is missing the page numbers 1726–1734; it also has an incorrect volume number of 31.  (That proceedings was published in seven volumes, so it's impossible for this article to be found in volume 31.)
