### 2025-08-27 #6468 @logological
@MathisFederico If you can't fix the figures, I suppose that's fine, but there are still a number of other issues from my list that haven't been properly addressed yet:

* Page 3: Change "leaderboard(Milani" to "leaderboard (Milani".
* Page 4: Change "more wide that deep" to "more wide than deep".
* Page 5: I suggested changing "require the further features extraction" to "require the further extraction of features".  You changed it instead to "require further features extraction", which is still grammatically incorrect.  To fix this, you could go with my originally proposed wording, or change your new text to "require further feature extraction".
* References: Wherever possible, please **provide DOIs and page numbers** for the bibliography items, and **check for and correct/remove any incorrect data** from them.  You fixed the two problems with Bacon et al. (2017) that I gave as examples, but did not correct any of the other items.

The DOI report by editorialbot above can be a good starting point to fix the DOIs in your references, but note that it's not necessarily correct or complete—for example, it failed to find DOIs for the arXiv papers, even though they definitely have DOIs.  But in these cases you might also consider citing the published version of the papers, where available, rather than the preprints.  For example, the PDDL paper was published in the proceedings of ICAPS 2020.
