### 2025-10-28 #6468 @MathisFederico
Thanks a lot @logological !
I'll synchronise with my teammates to finish the last DOIs now that you gave us a working example it will be much easier!
