### 2025-10-30 #6468 @logological
@MathisFederico: I see you've made some further changes to the references since accepting my pull request.  Unfortunately, these new changes have introduced several errors:

* You added `10.48550/arxiv.2002.06432` as the DOI for the `PDDLgym` paper, but this is the DOI for the preprint, not the published version listed in the BibTeX record.
* You added `10.48550/arxiv.2003.05012` as the DOI for the `milani2020minerl2019` paper, but once again this is the DOI for the preprint and not the published paper listed in the BibTeX record.
* You added `10.48550/arxiv.1510.09142` as the DOI for the `heess2016learning` paper, but once again this is the DOI for the preprint and not the published paper listed in the BibTeX record.
* You added `10.48550/arxiv.1805.08296` as the DOI for the `nachum2018data` paper, but once again this is the DOI for the preprint and not the published paper listed in the BibTeX record.

A DOI is intended to uniquely identify a published document, and so you should not use the DOI of one version of a document to refer to a different version of it that has different content and/or bibliographic metadata.

As far as I can tell, none of the four published papers above have DOIs, so could you please remove the `doi` fields from their respective BibTeX entries?
