### 2026-03-09 #6468 @MathisFederico
@logological sorry it has been a crazy early 2026, it completely slipped my mind.

Here is the version for including all changes: https://github.com/IRLL/HierarchyCraft/releases/tag/v2.0.1
Here is the Zenodo archive DOI: https://doi.org/10.5281/zenodo.18929144
