### 2026-03-10 #6468 @logological
@MathisFederico Thanks for this, but it seems you've reintroduced the same licence problem you had earlier.  Your Zenodo archive is tagged as "GNU General Public License v3.0 or later" but [you recently removed this same licence tag from `pyproject.toml`](https://github.com/IRLL/HierarchyCraft/commit/08109597cd7f8d05fbcfb79f614971aa694fc35f) in favour of "GNU General Public License v3 (GPLv3)".

Could you please either change the licence in your `pyproject.toml` and `README.md` to GPLv3+, or else change the licence in the Zenodo archive to GPLv3?
