### 2026-03-12 #6468 @logological
To the EiC team:  This note is to explain some anomalies in the review process and to argue that these should not present any barrier to the paper being accepted for publication.  First, a summary of the review process:

* I originally assigned two reviewers to this submission, @lwu9 and @Christopher-Henry-UM.

* @lwu9 completed all but three of her checklist items, then posted some questions about these three outstanding items.  Following a discussion with the author, she posted on 2024-05-27 that she was now satisfied on those three points.  However, she did not go back to update the checklist—I strongly suspect this was just absent-mindedness.  In the same message, she reported a couple error messages she encountered when running the software, but the author was unable to reproduce them and requested further troubleshooting information from her.  (Neither I, nor @Christopher-Henry-UM, nor the third reviewer I eventually brought on, reported any similar error messages.)  For unknown reasons, @lwu9 stopped responding to messages and e-mails after 2024-06-05.

* @Christopher-Henry-UM reviewed and signed off on every aspect of the submission except for Functionality and Performance. He left these items unchecked not because he believed the submission was deficient but because he determined in the course of his review he was not qualified to assess these two aspects.

* I therefore brought in a third reviewer, @inpefess, who reviewed all aspects of the paper. I checked the References box in his checklist for him, as he had apparently forgotten to do this; his last comment on 2025-03-10 explicitly states that he recommends publication after a minor correction to one of the references, which has long since been applied.

We're at the state now that each item in the reviewing checklist has been credibly signed off on by at least two reviewers.  Apart from @lwu9's error report, which she never followed up on and which the four others participating in this review did not observe or could not reproduce, there are no unresolved comments or questions about the submission.  I therefore believe that we can accept the submission on the basis of the reviewers' checklists and comments as a whole.  (At least, I don't see anything in the editorial policies that would prevent us from doing so, and would rather avoid having to bring in yet another reviewer on a submission that has already been under review for two years.)
