### 2026-03-14 #6468 @logological
@crvernon Could you confirm whether you've seen [my comment above](https://github.com/openjournals/joss-reviews/issues/6468#issuecomment-4050565156) regarding the checklists?
