### 2026-03-18 #6468 @MathisFederico
Done @crvernon , I change the Zenodo title to "HierarchyCraft: A Benchmark Builder for Hierarchical Reasoning" as the paper
