### 2026-03-19 #6468 @crvernon
🥳 Congratulations on your new publication @MathisFederico ! Many thanks to @logological for editing and @lwu9, @Christopher-Henry-UM, and @inpefess for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
